+++ Runtime/CommandLine/CmdArgNetStart.cs
@@ -1,0 +1,52 @@
+// Copyright (C) 2021-2024 Steffen Itterheim
+// Refer to included LICENSE file for terms and conditions.
+
+using CodeSmile.Components;
+using System;
+using UnityEngine;
+
+namespace CodeSmile.Netcode.CommandLine
+{
+	[Serializable]
+	public class CmdArgNetStart : CmdArgBase, ICmdArgImpl
+	{
+		public void Reset()
+		{
+			Description = "Starts the application as either server, host or client.";
+			Argument = "-netstart";
+			Parameters = new[] { "server", "host", "client" };
+		}
+
+		public void OnValidate()
+		{
+			if (Argument != null)
+				Argument = Argument.ToLower();
+		}
+
+		public async void Process()
+		{
+			if (Args.TryGetValue(Argument, out var mode))
+			{
+				var paramsIndex = 0;
+				if (mode.Equals(Parameters[paramsIndex++]))
+				{
+					Debug.Log("Starting Server ...");
+					SceneAutoLoader.DisableAll(); // server loads scene via NetworkSessionState
+					await NetworkStart.Server();
+				}
+				else if (mode.Equals(Parameters[paramsIndex++]))
+				{
+					Debug.Log("Starting Host ...");
+					SceneAutoLoader.DisableAll(); // server loads scene via NetworkSessionState
+					await NetworkStart.Host();
+				}
+				else if (mode.Equals(Parameters[paramsIndex++]))
+				{
+					Debug.Log("Starting Client ...");
+					SceneAutoLoader.DisableAll(); // clients auto-load when connected
+					await NetworkStart.Client();
+				}
+			}
+		}
+	}
+}